requirements.txt
numpy
PyYAML
pytest
tqdm
torch==2.10
huggingface-hub
kernels
setuptools
typing-extensions==4.15.0
datasets
tiktoken
sentencepiece
